# frontend/openapi.snapshot.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Job RAG API",
    "description": "RAG system for AI Engineer job postings \u2014 search, matching, and gap analysis",
    "version": "0.3.0"
  },
  "paths": {
...
    "/search": {
      "get": {
        "summary": "Search",
        "description": "Semantic search over job postings with optional RAG generation.\n\nWhen generate=True (default), returns an LLM-generated answer with sources.\nWhen generate=False, returns just the ranked search results.",
        "operationId": "search_search_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Q"
            }
          },
          {
            "name": "seniority",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Seniority"
            }
          },
          {
            "name": "remote",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Remote"
            }
          },
          {
            "name": "min_salary",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Salary"
            }
          },
          {
            "name": "generate",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Generate"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Search Search Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/match/{posting_id}": {
      "get": {
        "summary": "Match",
        "description": "Match a specific posting against the user profile.\n\n``posting_id`` is typed as ``uuid.UUID`` so malformed IDs are rejected\nwith 422 at the FastAPI boundary instead of surfacing as an asyncpg\ncast error (HTTP 500) inside the query.",
        "operationId": "match_match__posting_id__get",
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "AzureAD_PKCE_B2C_multi_tenant": []
          }
        ],
        "parameters": [
          {
            "name": "posting_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Posting Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Match Match  Posting Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/gaps": {
      "get": {
        "summary": "Gaps",
        "description": "Aggregate skill gaps across all (or filtered) postings.",
        "operationId": "gaps_gaps_get",
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "AzureAD_PKCE_B2C_multi_tenant": []
          }
        ],
        "parameters": [
          {
            "name": "seniority",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Seniority"
            }
          },
          {
            "name": "remote",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Remote"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Gaps Gaps Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/dashboard/top-skills": {
      "get": {
        "tags": [
          "dashboard"
        ],
        "summary": "Dashboard Top Skills",
        "description": "Top-N skills with must-have / nice-to-have split (DASH-01).\n\nSoft skills hidden by default (D-13); pass ``?include_soft=true`` to include.\nReturns ``DashboardTopSkillsResponse`` with named OpenAPI schema (drives\nopenapi-typescript codegen).",
        "operationId": "dashboard_top_skills_dashboard_top_skills_get",
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "AzureAD_PKCE_B2C_multi_tenant": []
          }
        ],
        "parameters": [
          {
            "name": "country",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/CountryFilter",
              "default": "WW"
            }
          },
          {
            "name": "seniority",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Seniority"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Seniority"
            }
          },
          {
            "name": "remote",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/RemoteFilter",
              "default": "any"
            }
          },
          {
            "name": "include_soft",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Include Soft"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardTopSkillsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/dashboard/salary-bands": {
      "get": {
        "tags": [
          "dashboard"
        ],
        "summary": "Dashboard Salary Bands",
        "description": "Salary percentiles p25/p50/p75 via PostgreSQL ``percentile_cont`` (DASH-02).\n\nReturns ``DashboardSalaryBandsResponse``; ``p25/p50/p75`` are ``int | None``\n(NULL when the filter matches zero salary-bearing postings - RESEARCH Pitfall 2).\nSalaries are stored as EUR/year (annualized at extraction time), so no\nper-period re-normalization happens here.",
        "operationId": "dashboard_salary_bands_dashboard_salary_bands_get",
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "AzureAD_PKCE_B2C_multi_tenant": []
          }
        ],
        "parameters": [
          {
            "name": "country",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/CountryFilter",
              "default": "WW"
            }
          },
          {
            "name": "seniority",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Seniority"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Seniority"
            }
          },
          {
            "name": "remote",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/RemoteFilter",
              "default": "any"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardSalaryBandsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/agent": {
      "post": {
        "summary": "Agent Query",
        "description": "Run the LangGraph agent to completion on a single query.",
        "operationId": "agent_query_agent_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentQuery"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Agent Query Agent Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "AgentQuery": {
        "properties": {
          "query": {
            "type": "string",
            "title": "Query"
          }
        },
        "type": "object",
        "required": [
          "query"
        ],
        "title": "AgentQuery"
      },
...
      "CountryFilter": {
        "type": "string",
        "enum": [
          "PL",
          "DE",
          "EU",
          "WW"
        ],
        "title": "CountryFilter",
        "description": "Canonical 4-value country filter (D-07).\n\nPL/DE: exact ISO-3166 alpha-2 match.\nEU:    union of EU-27 ISO codes OR location_region == 'EU' (catches NULL-country\n       \"Remote (EU)\" postings per Phase 2 D-09).\nWW:    no filter (default; matches Worldwide)."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Seniority": {
        "type": "string",
        "enum": [
          "junior",
          "mid",
          "senior",
          "staff",
          "lead",
          "unknown"
        ],
        "title": "Seniority"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      },
      "AzureAD_PKCE_B2C_multi_tenant": {
        "type": "oauth2",
        "description": "`Leave client_secret blank`",
        "flows": {
          "authorizationCode": {
            "scopes": {
              "api://00000000-0000-0000-0000-000000000000/access_as_user": "access_as_user"
            },
            "authorizationUrl": "https://login.microsoftonline.com/common/oauth2/v2.0/authorize",
            "tokenUrl": "https://login.microsoftonline.com/common/oauth2/v2.0/token"
          }
        }
      }
    }
  }
}